Update JS-SDK
In this commit we updated the js-sdk (wich fixed the bug in initEmailService, that
caused multiple calls of the function for different dom elements to bind to the last of them)

We also updated the way js-sdk is included. Its included through npm repository now, as a
npm dependency. This way we have controll over which version is used.

Finally we updated the webpack config and some other npm dependencies a bit.

Changes to be committed:
	modified:   endereco.js
	new file:   endereco.scss
	modified:   metadata.php
	modified:   out/assets/js/endereco.min.js
	deleted:    out/assets/js/oxid6-bundle.js
	modified:   package.json
	modified:   webpack.config.js