Add basic QA automatisation and update js-sdk
Given that the plugin should support a bunch of different versions of oxid shop, I've
added a basic QA automation, that checks if the code fits PSR12 standard, checks compatibility with
php 7.0 - 8.1 and compatibility with different shop versions.

Apart from that i've updated the js-sdk to the latest version and rebuild the bundle.