Change the load sequence for the resources
In some cases the ams is initiated for both the delivery and billing addresses
at the same time, which can lead to modals appearing at the same time.
The problem is in js-sdk, but as a quick solution, I would load shipping
ams with delay, which gives billing address enough time to render the modal.

Ref: O6M-63