Remove mandatory paypal express check
Oxid stores address data in the created paypal express user. Every time a user orders and using the same paypal account
his address is just copied into this user. Which can deprecate our statuscodes and make a new address check required, which
however will not be recognized by our module.
To fix it, we would clean paypal users statuscodes every time it was loaded in the checkout essentially triggering a new addresscheck.
But this lead to a situation, where the address can not be confirmed, as the statuscodes with confirmation mark are resetted every time the page loads.
As this second problem appears worse then the first, the quickfix would be to allow the statuscodes to persit between page loads.

Ref: O6M-67