Sign addresses with hash
Existing customer check and its variant paypal express checkout check have to reliably detect if an address has been changed. In
order to do that, the logic would check the status codes of the address. If they are empty or have the status "not_checked", then its
assumed that the address has to be checked (basically its assumed to be a new address).
However if the address is changed in some ways, that dont also change the statuses (e.g. when paypal express checkout imports data) then
its impossible to detect that the address need a new check, as the statuscodes from the last check are still there.
So in order to solve this problem, the address receives a hash signature, which is calculated based on the address data.
The signature is created right after the address has been checked, which ensures the validity of the status codes. When checking if the address
needs a new check, the hash is calculated anew and if the new hash and the saved hash are not the same, then the address has been
modified since last time, so it should be checked again.

Ref: O6M-67