Extend generation logic for MOJOISO31662 column
Currently our plugin assumes, that the oxstates column OXISOALPHA2 contains the second part
of the ISO 3166-2. Which we use to build and prefill our MOJOISO31662 column (it must contain
the ISO 3166-2). However we made experience, that some shops change that field to contain the full
code for some reason.
So in order to increase the robustness of our plugin, wenn check the format of the content of that
column, and if it fits the pattern of ISO 3166-2, then we just use the value. Otherwise, we try to
generate the code with old logic.

Ref: O6M-70

Co-authored-by: Jonas Joachim <jonas@endereco.de>