Make subdivision mapping more robust
Our original assumtion based on the default data in the oxstate table was, that the OXID is always a second part of
ISO 3166-2 code. Which turns out not to be the case.
So in order to be able to process a potentially random ID such as hash or in case of our customer, who reported the problem, a
full ISO 3166-2 code (used as ID in OXID column), we need to redo the mapping table.

The mapping from full ISO 6166-2 code to OXID is direct 1-to-1, the mapping from OXID requires the country code for context.

Ref: O6M-80

Co-authored-by: Jonas Joachim <jonas@endereco.de>